+++ DesignToolbox/DesignToolbox/MainView.swift
@@ -35,7 +35,6 @@ struct MainView: View {
                     Label("app_bottomBar_about_label", image: "info-fill")
                 }
                 .tag(2)
-        }
-        .oudsAccentColor(theme.colors.contentBrandPrimary)
+        }.oudsAccentColor(theme.colors.contentBrandPrimary)
     }
 }